[#618] Fix payload_len handling
Use size_t/ssize_t to store the message buffer sizes instead of
a uint16_t and int types.

+++ coap/er-coap-13/er-coap-13.c
@@ -1499,8 +1499,7 @@ coap_set_header_size(void *packet, uint32_t size)
 /*-----------------------------------------------------------------------------------*/
 /*- PAYLOAD -------------------------------------------------------------------------*/
 /*-----------------------------------------------------------------------------------*/
-int
-coap_get_payload(void *packet, const uint8_t **payload)
+size_t coap_get_payload(void *packet, const uint8_t **payload)
 {
   coap_packet_t *const coap_pkt = (coap_packet_t *) packet;
 
@@ -1513,8 +1512,7 @@ coap_get_payload(void *packet, const uint8_t **payload)
   }
 }
 
-int
-coap_set_payload(void *packet, const void *payload, size_t length)
+size_t coap_set_payload(void *packet, const void *payload, size_t length)
 {
   coap_packet_t *const coap_pkt = (coap_packet_t *) packet;
